+++ packages/eslint-plugin-template/tests/rules/prefer-self-closing-tags/cases.ts
@@ -34,6 +34,8 @@ export const valid: readonly (string | ValidTestCase<Options>)[] = [
     <p>Fallback content</p>
   </ng-content>`,
   { code: '<app-root></app-root>', filename: 'src/index.html' },
+  '<ng-container>&nbsp;</ng-container>',
+  '<my-component>  <!-- not empty -->  </ng-container>',
 ];
 
 export const invalid: readonly InvalidTestCase<MessageIds, Options>[] = [