+++ packages/angular-eslint/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 19.1.0 (2025-02-09)
+
+### 🚀 Features
+
+- **eslint-plugin-template:** add rule prefer-static-string-properties ([#2228](https://github.com/angular-eslint/angular-eslint/pull/2228))
+
+### ❤️ Thank You
+
+- Dave @reduckted
+
 ## 19.0.2 (2024-12-10)
 
 This was a version bump only for angular-eslint to align it with other projects, there were no code changes.